move conda-build exceptions
